Upload a zip of SDK with CI
We upload an artifact of the entire device app SDK that can be
directly used as LIBDIR for device apps. It includes all include
files, the libraries, the linker script, and some short documentation.

Retention of the artifact is right now just set to 2 days, so if you
want to make a proper release be sure to copy it.

Co-authored-by: Michael Cardell Widerkrantz <mc@tillitis.se>